Use chroot to modify the ipa-ramdisk
So far we've used a very unreliable hack to modify the ipa-ramdisk
and, for example, install new updated content.
We switch to a different more reliable hack using chroot, which
allows us to trigger native mechanisms (like dependency check)
when installing the updated packages.

To be able to run the installation we need to remove the rpm lock
to prevent the ``BDB0091 DB_VERSION_MISMATCH: Database environment
version mismatch``.
In a normal system, we could this in a different way, for example
running ``dnf clean dbcache``, but doing that in a chroot
environment inside a container will simply raise more problems
related to dnf.