Use dnf to install packages in the ipa ramdisk
Move the logic to modify the ipa-ramdisk to the chrooted.sh script
that is executed in a chroot environment.
This allows us to use native commands directly in the ipa-ramdisk
image, like dnf, and exit correctly in case of failures.
Also provide a list of packages installed in the ipa-ramdisk.