Unify packages list
Change name of default list of packages to install file to be
consistent with the other images.
Convert list of packages to install file in the Dockerfile to
ARG for easier customization.
Use ENV to define destination of list of packages to install in
only one place to force automated consistency between the
Dockerfile, and the build-ipa.sh and chrooted.sh scripts.